grid_trading_bot/requirements.txt
ccxt
PyYAML
numpy
pandas
ta  # Technical Analysis library
pytest
python-dotenv
sentry-sdk